further refactoring and fixing
