+++ cluster_tools/util.py
@@ -5,6 +5,7 @@ import os
 import threading
 import time
 import logging
+import re
 
 def local_filename(filename=""):
     return os.path.join(os.getenv("CFUT_DIR", ".cfut"), filename)
@@ -81,7 +82,7 @@ class FileWaitThread(threading.Thread):
     be created. When a specified file is created, it invokes a callback.
     """
 
-    def __init__(self, callback, interval=1):
+    def __init__(self, callback, executor, interval=1):
         """The callable ``callback`` will be invoked with value
         associated with the filename of each file that is created.
         ``interval`` specifies the polling rate.
@@ -92,6 +93,7 @@ class FileWaitThread(threading.Thread):
         self.waiting = {}
         self.lock = threading.Lock()
         self.shutdown = False
+        self.executor = executor
 
     def stop(self):
         """Stop the thread soon."""
@@ -124,37 +126,22 @@ class FileWaitThread(threading.Thread):
                         # Check for output file as a fast indicator for job completion
                         handle_completed_job(job_id, filename, False)
                     else:
-                        # If the output file was not found, we determine the job status so that
-                        # we can recognize jobs which failed hard (in this case, they don't produce output files)
-                        stdout, _, exit_code = call("scontrol show job {}".format(job_id))
+                        status = self.executor.check_for_crashed_job(job_id)
 
                         # We have to re-check for the output file since this could be created in the mean time
                         if os.path.exists(filename):
                             handle_completed_job(job_id, filename, False)
                         else:
-                            if exit_code != 0:
+                            if status == "completed":
                                 logging.error(
-                                    "Couldn't call scontrol to determine job's status. {}. Continuing to poll for output file. This could be an indicator for a failed job which was already cleaned up from the slurm db. If this is the case, the process will hang forever."
+                                    "Job state is completed, but {} couldn't be found.".format(
+                                        filename
+                                    )
                                 )
-                            else:
-                                job_state_search = re.search('JobState=([a-zA-Z_]*)', str(stdout))
-
-                                if job_state_search:
-                                    job_state = job_state_search.group(1)
-
-                                    if job_state in SLURM_STATES["Failure"]:
-                                        handle_completed_job(job_id, filename, True)
-                                    elif job_state in SLURM_STATES["Ignore"]:
-                                        # This job state can be ignored
-                                        pass
-                                    elif job_state in SLURM_STATES["Unclear"]:
-                                        logging.warn("The job state for {} is {}. It's unclear whether the job will recover. Will wait further".format(job_id, job_state))
-                                    elif job_state in SLURM_STATES["Success"]:
-                                        logging.error(
-                                            "Job state is completed, but {} couldn't be found.".format(
-                                                filename
-                                            )
-                                        )
-                                        handle_completed_job(job_id, filename, True)
+                                handle_completed_job(job_id, filename, True)
+                            elif status == "failed":
+                                handle_completed_job(job_id, filename, True)
+                            elif status == "ignore":
+                                pass
 
             time.sleep(self.interval)
